// src/app/core/models/pharaoh.model.ts
export interface Pharaoh {
  id: string;
  slug: string;
  name: string;
  throneName?: string;
  dynasty: string;
  dynastyNumber: number;
  reign: string;
  approximateDates: string;
  period: string;
  biography: string;
  achievements: string[];
  monuments: string[];
  family: string;
  importantEvents: string[];
  relatedArticles: string[];
  image: string;
  imageAlt: string;
}